Openweb implementation change #111 (#127)

+++ commons/src/main/scala/com/flipkart/connekt/commons/services/DeviceDetailsService.scala
@@ -175,7 +175,7 @@ object DeviceDetailsService extends Instrumented {
               override def run() = {
                 try {
                   DistributedCacheManager.getCache(DistributedCacheType.DeviceDetails).put[DeviceDetails](chunk.toList.map(d => (cacheKey(appName, d.deviceId), d)))
-                  warmUpTasks(jobId).currentCount.getAndAdd(chunk.size)
+                  warmUpTasks(jobId).currentCount.getAndAdd(chunk.size.toLong)
                 }
                 catch {
                   case e: Exception =>
