+++ commons/src/main/scala/com/flipkart/connekt/commons/services/ReportingService.scala
@@ -67,7 +67,7 @@ class ReportingService(reportManagerDao: StatsReportingDao) extends TService wit
   def recordPushStatsDelta(clientId: String, contextId: Option[String], stencilId: Option[String], platform: Option[String], appName: String, event: String, count: Int = 1): Unit = {
 
     val datePrefix = DateTimeUtils.calenderDate.print(Calendar.getInstance().getTimeInMillis) + "."
-    updateTagCounters(clientId, count, datePrefix, contextId.orNull, Channel.PUSH, stencilId.orNull)(platform.orNull, appName.toLowerCase, event)
+    updateTagCounters(clientId, count.toLong, datePrefix, contextId.orNull, Channel.PUSH, stencilId.orNull)(platform.orNull, appName.toLowerCase, event)
     contextId.foreach(id => updateLastSeen(s"$datePrefix${clientId}.$id"))
   }
